requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database (SQLite for development - no additional packages needed)
# psycopg2-binary==2.9.9  # Uncomment when PostgreSQL is available
# redis==5.0.1  # Uncomment when Redis is available
# django-redis==5.4.0  # Uncomment when Redis is available

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7
bcrypt==4.1.2

# File handling & Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
botocore==1.34.0

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Real-time features
channels==4.0.0
# channels-redis==4.1.0  # Uncomment when Redis is available
daphne==4.0.0

# API documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Third-party integrations
twilio==8.10.3
openai==1.3.7
stripe==7.8.0
razorpay==1.3.0
sendgrid==6.10.0

# Email handling
django-anymail==10.2
imapclient==2.3.1
email-validator==2.1.0

# Data processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# Development tools
django-debug-toolbar==4.2.0
factory-boy==3.3.0
faker==20.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Validation
marshmallow==3.20.1
django-filter==23.5 